To begin with, I'll make an option that http staging in happens as part of the job; with a mechanism to switch that between being part of the job and the former stage-as-separate-task mode.
This could actually be implemeted as a usable configurable feature allowing
AWS execution as long as HTTP/FTP is used for staging - so that it would
be an intermediate step to data management but actually usable as proof
of concept.

probably should start with some kind of linear sequence app



===

other notes, on potential patch series:


Add in multiple stage-in staging points

Prior to this commit, stage-in annotations happen as another parsl task,
either on the same executor as the annotated task or on the data manager's
executor.

This patch should add in other staging points such as:

* asking the executor to perform the stage-in
    - which might be condor/work_queue staging
    - or the executor might have a Channel it can use to access workdir

* more tightly binding to the job (so that the task turns into a
    sequenced task of stage in and execute)


Stage-out is deliberately omitted here because it needs some rearranging
to be specified this way.